updated android build files, updated ndk and other tool versions, fixed an exit... omg.. lets not discuss this..

+++ bash/android/gradle_pathing.sh
@@ -1,16 +1,25 @@
 #!/bin/bash
 
+# global variables
+is_sdl_image=true
+is_sdl_mixer=true
 echo "  > gradle pathing check"
 pathing_locker_id=0
 pathing_locker="/tmp/zox_gradle_building.lock"
 
 clear_gradle_build() {
     echo "  > clearing gradle pathing lock"
-    rm "$pathing_locker"
+    if [ -f "$pathing_locker" ]; then
+        rm "$pathing_locker"
+    fi
 }
 
 # Function to start the Gradle build
 start_gradle_build() {
+    if [ -z "$sdl_directory" ]; then
+        echo " > clearing gradle as failed to clear last run"
+        clear_gradle_build
+    fi
     # Check if the lock file exists
     if [ -f "$pathing_locker" ]; then
         return 1
@@ -18,40 +27,57 @@ start_gradle_build() {
     # Create the lock file
     pathing_locker_id=$PPID
     touch "$pathing_locker"
+    # global variables
+    ndk_ver=25 # 15c / 13b / 21 / 26b
+    jdk_ver=20 # 7 / 11 / 20
+    platform_ver=34 # 30
+    platform_tools_ver=34.0.5
+    build_tools_ver=34.0.0-rc4 # 30.0.3
     echo "  > gradle setting paths"
     zoxel_directory=$PWD
     android_sdk_path="$zoxel_directory/build/android_sdk"
-    ndk_path="$android_sdk_path/ndk"
+    apksigner=$android_sdk_path/build-tools/$build_tools_ver/apksigner
+    ndk_path=$android_sdk_path/ndk
     export ANDROID_HOME=$android_sdk_path
     export ANDROID_SDK_ROOT=$android_sdk_path
     export ANDROID_NDK_HOME=$ndk_path
     # export zoxel_directory="$zoxel_directory"
-    android_directory="$zoxel_directory/build/android-build"
-    android_bash_directory="$zoxel_directory/bash/android"
+    android_directory=$zoxel_directory/build/android-build
+    android_build_directory=$zoxel_directory/build/android-build
+    android_bash_directory=$zoxel_directory/bash/android
     # export JAVA_HOME="/usr/lib/jvm/java-17-openjdk/bin/java"
     # Get the result of 'whereis jdk'
-    jdk_whereis=$(whereis java) # locates jdk through whereis
-    jdk_whereis2=$(echo $jdk_whereis | awk '{print $NF}') # gets the last string in the list
-    jdk_whereis2=$(dirname "$(dirname "$jdk_whereis2")") # removes last 2 directories
-    JAVA_HOME=$jdk_whereis2 #"/usr/lib/jvm/java-17-openjdk"
+    # jdk_whereis=$(whereis java) # locates jdk through whereis
+    # jdk_whereis2=$(echo $jdk_whereis | awk '{print $NF}') # gets the last string in the list
+    # jdk_whereis2=$(dirname "$(dirname "$jdk_whereis2")")
+    # JAVA_HOME=$jdk_whereis2 #"/usr/lib/jvm/java-17-openjdk"
+    JAVA_HOME=/usr/lib/jvm/java-$jdk_ver-openjdk # 7 / 11 / 20
+    # javaHome=$JAVA_HOME
     export JAVA_HOME=$JAVA_HOME # "/usr/lib/jvm/java-17-openjdk" # $(dirname $(dirname $(whereis java)))
     # SDL
-    sdl_directory="$zoxel_directory/build/sdl/sdl" # "$HOME/SDL/SDL2"
-    export sdl_directory="$zoxel_directory/build/sdl/sdl" # "$HOME/SDL/SDL2"
-    sdl_android_project_directory="$sdl_directory/android-project"
-    sdl_image_directory="$zoxel_directory/build/sdl/sdl_image"  # $HOME/SDL/SDL2_image
-    sdl_mixer_directory="$zoxel_directory/build/sdl/sdl_mixer"  # $HOME/SDL/SDL2_mixer
+    sdl_parent_directory=$zoxel_directory/build/sdl
+    sdl_directory=$sdl_parent_directory/sdl
+    export sdl_directory=$sdl_parent_directory/sdl
+    sdl_android_project_directory=$sdl_directory/android-project
+    sdl_image_directory=$sdl_parent_directory/sdl_image
+    sdl_mixer_directory=$sdl_parent_directory/sdl_mixer
+    # tools
+    apk_filepath="build/android-build/app/build/outputs/apk/release/zoxel.apk"
+    adb=$android_sdk_path/platform-tools/adb
+    package_name="org.libsdl.app"
+    activity_name="SDLActivity"
+    log_tag="Zoxel"
     # print paths
     echo "      + zoxel_directory is $zoxel_directory"
-    echo "      + android sdk path to [$ANDROID_SDK_ROOT]"
-    echo "      + ndk path to [$ANDROID_NDK_HOME]"
-    echo "      + java path is [$JAVA_HOME]"
     echo "      + android_directory [$android_directory]"
     echo "      + android_bash_directory [$android_bash_directory]"
     echo "      + sdl_android_project_directory [$sdl_android_project_directory]"
     echo "      + sdl_directory [$sdl_directory]"
     echo "      + sdl_image_directory [$sdl_image_directory]"
     echo "      + sdl_mixer_directory [$sdl_mixer_directory]"
+    echo "      + android sdk path to [$ANDROID_SDK_ROOT]"
+    echo "      + ndk path to [$ANDROID_NDK_HOME]"
+    echo "      + java path is [$JAVA_HOME]"
     return 0
 }
 
@@ -88,3 +114,5 @@ end_gradle_build() {
 # export ANDROID_HOME="$android_sdk_directory/sdk" # "/usr/lib/android-sdk"
 # ANDROID_SDK_ROOT="$ANDROID_HOME" # "/usr/lib/android-sdk"
 # export ANDROID_SDK_ROOT="$ANDROID_HOME" #"/usr/lib/android-sdk"
+
+# $ANDROID_SDK_ROOT/platform-tools/adb
